requirements.txt
fastapi==0.111.0
pydantic_settings==2.3.1
asyncpg==0.29.0
sqlalchemy==2.0.30
filetype==1.2.0
loguru==0.7.2
alembic==1.13.1
